Trait r3_port_riscv::TimerOptions

pub trait TimerOptions {
    const MTIME_PTR: usize;
    const MTIMECMP_PTR: usize;
    const RESET_MTIME: bool;
    const FREQUENCY: u64;
    const FREQUENCY_DENOMINATOR: u64;
    const HEADROOM: u32;
    const INTERRUPT_NUM: InterruptNum;
}

The options for use_timer!.

Associated Constants

const MTIME_PTR: usize

The memory address of the mtime register.

const MTIMECMP_PTR: usize

The memory address of the mtimecmp register.

const RESET_MTIME: bool

When set to true, the driver clears the lower 32 bits of the mtime register on boot.

Disabling this might increase the runtime overhead of the driver. Nevertheless, the need to disable this might arise for numerous reasons including:

Updating the mtime register is not supported by QEMU at this time.

The mtime register might be shared with other harts and clearing it could confuse the code running in the other harts.

const FREQUENCY: u64

The numerator of the effective timer clock rate of the dual timer.

const FREQUENCY_DENOMINATOR: u64

The denominator of the effective timer clock rate of the dual timer. Defaults to 1.

const HEADROOM: u32

The maximum permissible timer interrupt latency, measured in hardware timer cycles.

Defaults to min(FREQUENCY * 60 / FREQUENCY_DENOMINATOR, 0x40000000).

const INTERRUPT_NUM: InterruptNum

The timer’s interrupt number. Defaults to INTERRUPT_TIMER.

Loading content...